Phew! This is a big bucket of PHP but it does a nice task for us.
Basically - we need one of our components to display the tournament in its entirety - meaning we do a huge data query and return the object back to our frontend.

I'm hoping this is where vue will shine to be honest.
On the frontend I plan to loop over this element, and pass the ID's of each match and each team on a 'filter down' basis.

Meaning less, but efficient code. I hope. Don't quote me guys.